Move files into bot or api folder
