+++ README.md
@@ -11,16 +11,16 @@ pip install -r requirements.txt
 
 ### API
 1. Set up a `.env` file in the API folder. See the `.env.example` file for reference.
-2. Run `API.py` to start the API:
+2. Run `api.py` to start the API:
 ```bash
-uvicorn API:app
+uvicorn api:app
 ```
 
 ### Bot
 1. Create a new Discord bot in the [Developer Portal](https://discord.com/developers/applications).
 2. Invite the bot to your server with the "application.commands" permission.
 3. Set up a `.env` file. See the `.env.example` file for reference. Add your Discord Bot Token from the Developer Portal as `BOT_TOKEN`.
-4. Add your API URL to the `bot_config.json` file as `API_URL` (if running locally, this is likely `http://127.0.0.1:8000`). Customise the other settings if you wish, or leave them as defaults.
+4. Add your API URL to the `config.json` file as `API_URL` (if running locally, this is likely `http://127.0.0.1:8000`). Customise the other settings if you wish, or leave them as defaults.
 5. Run `pint_bot.py` to start the bot:
 ```bash
 python pint_bot.py
@@ -47,10 +47,10 @@ pytest
 
 ## Customising
 - Customise the API by modifying the `.env` file in the API folder (e.g., how much debt to allow per transaction, the API endpoint names).
-- Customise the bot by modifying the `bot_config.json` file.
+- Customise the bot by modifying the `config.json` file.
 - Customise the way the bot sends messages in the `send_messages.py` file (by default it uses embeds).
 
-## bot_config.json
+## config.json
 - **BOT_NAME**: The name of the bot in Discord.
 - **API_URL**: The URL of your API for the economy.
 - **CURRENCY_NAME**: The name of the currency used in the economy.
