+++ bot/.env.example
