+++ bot/bot_commands.py
@@ -1,11 +1,11 @@
 from fractions import Fraction
 import discord 
 from discord import app_commands
-from API.models import OweRequest, SetUnicodePreferenceRequest, SettleRequest
+from api.models import OweRequest, SetUnicodePreferenceRequest, SettleRequest
 import api_client
 from error_handling import handle_error
 from formatter import currency_formatter, to_percentage
-from send_messages import send_info_message, send_one_column_table_message, send_success_message, send_two_column_table_message
+from bot.send_messages import send_info_message, send_one_column_table_message, send_success_message, send_two_column_table_message
 
 async def fetch_unicode_preference(interaction, user_id) -> bool:
     """Fetch the user's Unicode preference from the API."""
