+++ bot/config.json
