+++ bot/formatter.py
