+++ tests/test_error_handling.py
@@ -29,7 +29,7 @@ class TestGetErrorMessage:
         assert "TestBot" in result["title"]
         assert "Doesn't Like Your Data" in result["title"]
 
-    @patch("bot_config.get_config")
+    @patch("config.get_config")
     def test_get_error_message_unknown(self, mock_get_config, mock_config):
         mock_get_config.return_value = mock_config
         result = get_error_message("NON_EXISTENT_CODE")
